Normalize recommended Git versions chart
This adds support for a new option that automatically normalizes a
stacked area chart. With normalization turned on, the area chart will
render the datasets with relative values and adjust the y axis to show
percentages instead of absolute values.

This is interesting for the recommended Git versions chart, where it
makes sense to show the relative number of recommended clients in order
to track upgrade trends better.